requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
datasets==3.0.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.6.1
gdown==5.2.0
h11==0.14.0
haystack-ai==2.5.1
haystack-experimental==0.1.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
idna==3.9
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
lazy-imports==0.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-plain==1.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.45.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
posthog==3.6.5
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
pypdf==4.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.2
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wikipedia==1.4.0
xxhash==3.5.0
yarl==1.11.1
